Move listFields to utils.dart
